# app/src/main/resources/openapi.yaml
openapi: 3.1.0
...
info:
  title: StudyDeck AI API
  version: 1.0.0
  summary: API versionada para una plataforma de estudio tipo Anki con IA, RAG, MCP y observabilidad.
  description: |
    OpenAPI 3.1 para StudyDeck AI.
    Diseño orientado a monolito modular hexagonal en Java 25 + Spring Boot 4.x + Spring AI 2.x.
    Defaults asumidos:
      - API versionada en /v1
      - paginación por offset: page=0, size=20, maxSize=100
      - sorting por createdAt asc salvo endpoints de revisión
      - autenticación principal: OAuth2/OIDC + JWT; API keys con scopes para automatización cuando aplique
      - roles y scopes descritos en securitySchemes y descripciones de operación
servers:
  - url: http://localhost:8080
    description: Backend local en Docker Compose
  - url: http://localhost:8080/actuator
    description: Actuator local en Docker Compose
  - url: https://api.studydeck.ai
    description: Producción
...
security:
  - oauth2:
      - study.read
paths:
  /v1/account/export:
    get:
      tags: [Account]
      summary: Export caller's own data (GDPR data portability)
      operationId: exportAccount
      security:
        - oauth2: [export.read]
      responses:
        "200":
          description: Full account data export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountExport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/account:
    delete:
      tags: [Account]
      summary: Permanently delete caller's account and all associated data (GDPR right to erasure)
      operationId: deleteAccount
      security:
        - oauth2: [study.write]
      responses:
        "204":
          description: Account and all data permanently deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      description: |
        OAuth2/OIDC para JWT bearer tokens.
        Vite/React SPA: authorizationCode + PKCE contra el IdP.
        Service-to-service y automatización MCP: clientCredentials con scopes del IdP.
        CLI interactivo: Device Authorization Grant documentado en el IdP.
        API keys son una alternativa machine-readable separada para automatización/MCP cuando no hay identidad de usuario final.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.studydeck.ai/oauth2/authorize
          tokenUrl: https://auth.studydeck.ai/oauth2/token
          refreshUrl: https://auth.studydeck.ai/oauth2/token
          scopes:
            openid: Identidad OIDC
            profile: Perfil del usuario
            email: Email del usuario
            study.read: Lectura de estudio
            study.write: Escritura de estudio
            review.write: Revisión
            documents.read: Lectura de documentos
            documents.write: Escritura de documentos
            rag.query: Consultas RAG
            ai.generate: Generación IA
            import.write: Importación JSON
            export.read: Exportación
            mcp.invoke: Invocar MCP
            admin.read: Lectura administrativa
            admin.write: Escritura administrativa
            audit.read: Lectura de auditoría
            observability.read: Lectura de observabilidad
        clientCredentials:
          tokenUrl: https://auth.studydeck.ai/oauth2/token
          scopes:
            study.read: Lectura de estudio
            study.write: Escritura de estudio
            documents.read: Lectura de documentos
            documents.write: Escritura de documentos
            rag.query: Consultas RAG
            ai.generate: Generación IA
...
  responses:
...
    Unauthorized:
      description: No autenticado
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: No autorizado
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    AccountExport:
      type: object
      required: [account, exportedAt]
      properties:
        account:
          type: object
          required: [id, createdAt]
          properties:
            id:
              type: string
              format: uuid
            subject:
              type: string
            email:
              type: string
              format: email
            displayName:
              type: string
            createdAt:
              type: string
              format: date-time
        decks:
          type: array
          description: Per-deck flashcard export payloads (same shape as FlashcardImportV1)
          items:
            $ref: "#/components/schemas/FlashcardImportV1"
        documents:
          type: array
          items:
            type: object
            required: [id, title, sourceType, ingestStatus, createdAt]
            properties:
              id:
                type: string
                format: uuid
              title:
                type: string
              sourceType:
                type: string
              ingestStatus:
                type: string
              createdAt:
                type: string
                format: date-time
        exportedAt:
          type: string
          format: date-time
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri-reference
          default: about:blank
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
        requestId:
          type: string
        traceId:
          type: string
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
    Violation:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
        code:
          type: string
...
    MultipleChoiceOption:
      type: object
      required: [key, text]
      properties:
        key:
          type: string
          pattern: "^[A-Z]$"
        text:
          type: string
          minLength: 1
          maxLength: 1000
...
    FlashcardImportV1:
      $schema: https://json-schema.org/draft/2020-12/schema
      type: object
      additionalProperties: false
      required: [schemaVersion, deck, notes]
      properties:
        schemaVersion:
          type: string
          const: "1.0"
        deck:
          type: object
          additionalProperties: false
          required: [title]
          properties:
            title:
              type: string
              minLength: 1
              maxLength: 120
            description:
              type: string
              maxLength: 1000
            tags:
              type: array
              items:
                type: string
        notes:
          type: array
          minItems: 1
          maxItems: 5000
          items:
            $ref: "#/components/schemas/ImportNote"
      examples:
        - schemaVersion: "1.0"
          deck:
            title: "Ejemplo"
          notes:
            - noteType: "basic"
              front: "Pregunta"
              back: "Respuesta"
    ImportNote:
      oneOf:
        - $ref: "#/components/schemas/ImportBasicNote"
        - $ref: "#/components/schemas/ImportReversedNote"
        - $ref: "#/components/schemas/ImportClozeNote"
        - $ref: "#/components/schemas/ImportMultipleChoiceNote"
        - $ref: "#/components/schemas/ImportFreeTextNote"
    ImportCommon:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
        source:
          type: object
          additionalProperties: false
          properties:
            type:
              type: string
            reference:
              type: string
    ImportBasicNote:
      allOf:
        - $ref: "#/components/schemas/ImportCommon"
        - type: object
          additionalProperties: false
          required: [noteType, front, back]
          properties:
            noteType:
              const: basic
            front:
              type: string
            back:
              type: string
    ImportReversedNote:
      allOf:
        - $ref: "#/components/schemas/ImportCommon"
        - type: object
          additionalProperties: false
          required: [noteType, front, back]
          properties:
            noteType:
              const: reversed
            front:
              type: string
            back:
              type: string
    ImportClozeNote:
      allOf:
        - $ref: "#/components/schemas/ImportCommon"
        - type: object
          additionalProperties: false
          required: [noteType, text]
          properties:
            noteType:
              const: cloze
            text:
              type: string
              pattern: ".*\\{\\{c[0-9]+::.+\\}\\}.*"
    ImportMultipleChoiceNote:
      allOf:
        - $ref: "#/components/schemas/ImportCommon"
        - type: object
          additionalProperties: false
          required: [noteType, question, options, correctOptionKeys]
          properties:
            noteType:
              const: multiple-choice
            question:
              type: string
            options:
              type: array
              minItems: 4
              maxItems: 5
              description: Multiple Choice P0 permite exactamente variantes ABCD o ABCDE.
              items:
                $ref: "#/components/schemas/MultipleChoiceOption"
            correctOptionKeys:
              type: array
              minItems: 1
              maxItems: 1
              uniqueItems: true
              description: Debe contener exactamente una opción correcta.
              items:
                type: string
                pattern: "^[A-Z]$"
            explanation:
              type: string
    ImportFreeTextNote:
      allOf:
        - $ref: "#/components/schemas/ImportCommon"
        - type: object
          additionalProperties: false
          required: [noteType, prompt, expectedAnswer]
          properties:
            noteType:
              const: free-text
            prompt:
              type: string
            expectedAnswer:
              type: string
            gradingGuidance:
              type: string
...